Fix authentication - use username + token for cache
This fix allows the use of ES with username + token authentication.
The plugin does authentication in the listProjectsFor function, using
the given token to get a list of projects for the given token.  If token
auth fails, updateCache will not add the user to the cache, and will not
proceed with the request.  The cache is changed to use the username and
the token as the key, so that the cache will report failure if a username
is given with a bogus key.